for SCCM and Intune compatibility

+++ silent-tester-runner-windows.ps1
@@ -1,31 +1,67 @@
-$TenantID = "TENANT_ID" # Replace TENANT_ID with you actual Microsoft Tenant Id
-$TestID = "TEST_ID" # Replace with TEST_ID Which Must be different for each test
+[CmdletBinding()]
+param (
+  [Parameter(Mandatory=$false)]
+  [string] $TestID = "TEST_ID", # TEST_ID must be different for each test, unless you use the AllowMultipleRuns switch
+  [Parameter(Mandatory=$false)]
+  [int] $ScenarioDuration = 86400, # defaults to 24 hours
+  [Parameter(Mandatory=$false)]
+  [string] $TenantID = "TENANT_ID", # We recommend replacing TENANT_ID in this script with your actual Microsoft Tenant Id
+  [Parameter(Mandatory=$false)]
+  [Alias ("SCCM","Intune")]
+  [switch] $UEM_Compatible_Mode, # Use for better compatibility with UEM solutions
+  [Parameter(Mandatory=$false)]
+  [Alias ("Force")]
+  [switch] $AllowMultipleRuns, # Use if you want to be able to run more the once with the same $TestID on the machine. 2 tabs might jump to user.
+  [Parameter(Mandatory=$false)]
+  [string] $CustomChromePath = "" # Use if you want to specify the path to the Chrome executable
+)
+#############
+### SETUP ###
+#############
+
+### Setting up the variables ###
 $pageURL = "https://st-sdk.ecdn.teams.microsoft.com/?customerId=${TenantID}&adapterId=PowerShell"
-$scenarioDuration = 86400 # defaults to 24 hours - can be changed
-$runOnce = $true # Set to $false if you want to be able to run more the one time on the machine. 2 tabs might jump to user.
-$customChromePath = ""
 $logPath = "$env:TEMP\p5_log_" + $TestID + ".txt"
 $errLogPath = "$env:TEMP\p5_err_" + $TestID + ".txt"
 $p5UserDataDir = "$env:TEMP\p5-user-" + $TestID
 $preferencesFilePath = $p5UserDataDir + "\Default\Preferences"
 $cacheFolderPath = $p5UserDataDir + "\Default\Cache"
 $defaultPaths = @("C:\Program Files (x86)\Microsoft\Edge\Application\msedge.exe","C:\Program Files (x86)\Google\Chrome\Application\chrome.exe", "C:\Program Files\Google\Chrome\Application\chrome.exe")
+$durationMinimum = 60
 
-$definition = @"
-      [DllImport("user32.dll")]
-      [return: MarshalAs(UnmanagedType.Bool)]
-      static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
-      public static void Hide(IntPtr hWnd) {
-        if ((int)hWnd > 0)
-            ShowWindow(hWnd, 0);
-      }
-"@
-add-type -MemberDefinition $definition -Namespace my -Name WinApi
-if ((Test-Path $logPath) -and $runOnce) {
-  Write-Host "Test $TestID already ran on this machine. aborting"
-  Exit
+### Parameter validation ###
+$RegexForTenantId = '[a-z0-9]{8}\-[a-z0-9]{4}\-[a-z0-9]{4}\-[a-z0-9]{4}\-[a-z0-9]{12}'
+if ($TenantID -notmatch $RegexForTenantId) {
+  Write-Error "Invalid Parameter: Tenant ID. Please provide a valid Tenant ID."
+  Exit 1
 }
-if (!$customChromePath -or !(Test-Path $customChromePath)) {
+if ($ScenarioDuration -lt $durationMinimum) {
+  Write-Error "Invalid Parameter: Scenario Duration. Please provide a Scenario Duration of greater than $durationMinimum seconds."
+  Exit 1
+}
+if ((Test-Path $logPath) -and (!$AllowMultipleRuns)) {
+  Write-Error "Test '$TestID' already ran on this machine. aborting"
+  Exit 1
+}
+
+### Function to hide the browser window ###
+$definition = @"
+  [DllImport("user32.dll")]
+  [return: MarshalAs(UnmanagedType.Bool)]
+  static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
+  public static void Hide(IntPtr hWnd) {
+    if ((int)hWnd > 0)
+      ShowWindow(hWnd, 0);
+  }
+"@
+Add-Type -MemberDefinition $definition -Namespace my -Name WinApi
+
+###################
+### MAIN SCRIPT ###
+###################
+
+### Selecting the Chrome executable path ###
+if (!$CustomChromePath -or !(Test-Path $CustomChromePath)) {
   try {
     if (Test-Path 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths\msedge.exe') {
       $edgeExe = Get-ItemPropertyValue 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths\msedge.exe' "(default)"
@@ -35,36 +71,44 @@ if (!$customChromePath -or !(Test-Path $customChromePath)) {
     # do nothing
   }
   if ($edgeExe) {
-    $customChromePath = $edgeExe
+    $CustomChromePath = $edgeExe
   }
   else {
     for($i = 0; $i -le $defaultPaths.Count; $i++ ) {
       if ($defaultPaths[$i] -and (Test-Path $defaultPaths[$i])) {
-        $customChromePath = $defaultPaths[$i]
+        $CustomChromePath = $defaultPaths[$i]
         break
       }
     }
   }
-  if (!$customChromePath) {
-    Write-Host "Could not find Edge or Chrome executable (chrome.exe), Please set `$customChromePath variable in the script to your Chrome executable path"
-    Exit
+  if (!$CustomChromePath) {
+    Write-Host "Could not find Edge or Chrome executable (chrome.exe), Please set `$CustomChromePath variable in the script to your Chrome executable path"
+    Exit 2
   }
 }
-Write-Host "Found Chrome path $($customChromePath)"
-$Process = Start-Process -RedirectStandardOutput $logPath -RedirectStandardError $errLogPath -passthru $customChromePath -ArgumentList "$($pageURL)  --hide-crash-restore-bubble --autoplay-policy=no-user-gesture-required --disable-backgrounding-occluded-windows --disable-background-media-suspend --disable-renderer-backgrounding --disable-gpu --remote-debugging-port=0 --disable-infobars --disable-restore-session-state --user-data-dir=$p5UserDataDir --disable-gesture-requirement-for-media-playback --disable-background-networking --disable-background-timer-throttling --disable-breakpad --disable-client-side-phishing-detection --disable-default-apps --disable-dev-shm-usage --disable-extensions --disable-field-trial-config --disable-features=site-per-process,WebRtcHideLocalIpsWithMdns --disable-hang-monitor --disable-popup-blocking --disable-prompt-on-repost --disable-sync --disable-translate --metrics-recording-only --no-first-run --safebrowsing-disable-auto-update --enable-automation --password-store=basic --use-mock-keychain --mute-audio --process-per-site" -WorkingDirectory $env:TEMP
+Write-Host "Found Chrome path '$CustomChromePath'"
+
+### Starting the browser process ###
+$Process = Start-Process -RedirectStandardOutput $logPath -RedirectStandardError $errLogPath -passthru $CustomChromePath -ArgumentList "$($pageURL)  --hide-crash-restore-bubble --autoplay-policy=no-user-gesture-required --disable-backgrounding-occluded-windows --disable-background-media-suspend --disable-renderer-backgrounding --disable-gpu --remote-debugging-port=0 --disable-infobars --disable-restore-session-state --user-data-dir=$p5UserDataDir --disable-gesture-requirement-for-media-playback --disable-background-networking --disable-background-timer-throttling --disable-breakpad --disable-client-side-phishing-detection --disable-default-apps --disable-dev-shm-usage --disable-extensions --disable-field-trial-config --disable-features=site-per-process,WebRtcHideLocalIpsWithMdns --disable-hang-monitor --disable-popup-blocking --disable-prompt-on-repost --disable-sync --disable-translate --metrics-recording-only --no-first-run --safebrowsing-disable-auto-update --enable-automation --password-store=basic --use-mock-keychain --mute-audio --process-per-site" -WorkingDirectory $env:TEMP
 if ([System.Security.Principal.WindowsIdentity]::GetCurrent().Name -ne "NT AUTHORITY\SYSTEM") {
   While ($Process.MainWindowHandle -eq 0) { Start-Sleep -m 100 }
   [my.WinApi]::Hide($Process.MainWindowHandle)
 }
-Write-Host "Started Chromium process, with id: $($Process.id)"
+Write-Host "$(Get-Date) Started Chromium process, with id: $($Process.id)"
+
+### Setting up the watchdog process ###
 $chromePid = $Process.id
 $cmd = "cmd.exe"
-$extraTimeout = $scenarioDuration + 10
-$argos =  "/c timeout ${extraTimeout} && taskkill.exe /f /t /pid ${chromePid}"
+$extraTimeout = $ScenarioDuration + 10
+$argos =  "/c timeout ${extraTimeout} && taskkill.exe /f /t /pid $chromePid && rd /s /q $cacheFolderPath"
 $watchdogProcess = Start-Process -WindowStyle hidden -passthru $cmd -ArgumentList $argos
-Start-Sleep -s $scenarioDuration
+if ($UEM_Compatible_Mode) {
+  return
+}
+
+### Waiting for the scenario to end and clean-up ###
+Start-Sleep -s $ScenarioDuration
 $stopProcessInfo = Stop-Process -InputObject $Process -passthru
-$stopProcessInfo
 if (Test-Path $preferencesFilePath) {
   try {
     $Prefs = ((Get-Content $preferencesFilePath) -replace "`"exit_type`":`"Crashed`"" , "`"exit_type`":`"none`"") -replace "`"exited_cleanly`":false","`"exited_cleanly`":true"
@@ -77,4 +121,4 @@ if (Test-Path $cacheFolderPath) {
     Remove-Item -Recurse $cacheFolderPath
   } catch {}
 }
-Write-Host "Stopped Chromium process"
+Write-Host "$(Get-Date) Stopped Chromium process"
